Move methods out to `routines.py`
Removing more code from `da.core`
